added textfile

+++ agencia de viagens/agencia de viagens/Node.h
@@ -2,7 +2,7 @@
 
 #include <string>
 #include "Accommodation.h"
-typedef long long int BigInt;
+typedef long int BigInt;
 using namespace std;
 class Node
 {
